pyproject.toml
[tool.poetry]
name = "projectmanager"
version = "0.1.0"
description = ""
authors = ["maxx <maxpetrikin@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.2.4"
python-dotenv = "^1.1.1"
djangorestframework = "^3.16.0"
psycopg2-binary = "^2.9.10"
drf-yasg = "^1.21.10"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
